+++ src/cli/add_remove.cpp
@@ -65,6 +65,8 @@ int image_add(const image_add_args& args, const global_settings& settings) {
     //
     // parse the cli args
     //
+
+    // the label is the label that will be given to the uenv added to the repo
     auto label = uenv::parse_uenv_label(args.label);
     if (!label) {
         term::error("the label {} is not valid: {}", args.label,
@@ -78,7 +80,7 @@ int image_add(const image_add_args& args, const global_settings& settings) {
         return 1;
     }
 
-    // parse input as either a label or a file path
+    // the source can be either the path of a squashfs file, or a label
     uenv_description source;
     if (const auto parse = parse_uenv_description(args.source); !parse) {
         term::error("invalid uenv specification: {}", parse.error().message());
@@ -87,8 +89,8 @@ int image_add(const image_add_args& args, const global_settings& settings) {
         source = parse.value();
     }
 
-    auto env = resolve_uenv(source, settings.config.repo,
-                            settings.calling_environment);
+    // derive the full description of the source uenv.
+    auto env = resolve_uenv(source, settings.config.repos);
     if (!env) {
         term::error("{}", env.error());
         return 1;
@@ -114,13 +116,17 @@ int image_add(const image_add_args& args, const global_settings& settings) {
     //
     // Open the repository
     //
-    if (!settings.config.repo) {
-        term::error("a repo needs to be provided either using the --repo "
-                    "option, or in the config file");
+
+    // the target repository is always the first repo in the list, which can be
+    // overriden using the --repo flag.
+    const auto repo = settings.config.repo();
+    if (!repo) {
+        term::error(
+            "a repo needs to be provided either using the --repo flag or "
+            "in the config file");
         return 1;
     }
-    auto store = uenv::open_repository(settings.config.repo.value(),
-                                       repo_mode::readwrite);
+    auto store = uenv::open_repository(repo->path, repo_mode::readwrite);
     if (!store) {
         term::error("unable to open repo: {}", store.error());
         return 1;
@@ -172,8 +178,7 @@ int image_add(const image_add_args& args, const global_settings& settings) {
     const auto uenv_paths = store->uenv_paths(sqfs->hash);
     uenv::uenv_date date{*util::file_creation_date(sqfs->sqfs)};
 
-    bool source_in_repo =
-        util::is_child(sqfs->sqfs, settings.config.repo.value());
+    bool source_in_repo = util::is_child(sqfs->sqfs, repo->path);
 
     // If an sqfs file is already in repo, and it was pulled from a repository
     // then there is a digest mismatch. Do not try to add this image
@@ -284,13 +289,13 @@ int image_rm([[maybe_unused]] const image_rm_args& args,
     spdlog::info("image rm {}", args.label);
 
     // open the repo
-    if (!settings.config.repo) {
+    const auto repo = settings.config.repo();
+    if (!repo) {
         term::error("a repo needs to be provided either using the --repo flag "
                     "in the config file");
         return 1;
     }
-    auto store = uenv::open_repository(settings.config.repo.value(),
-                                       repo_mode::readwrite);
+    auto store = uenv::open_repository(repo->path, repo_mode::readwrite);
     if (!store) {
         term::error("unable to open repo: {}", store.error());
         return 1;