+++ src/cli/pull.cpp
@@ -52,8 +52,7 @@ void image_pull_args::add_cli(CLI::App& cli,
     pull_cli->footer(image_pull_footer);
 }
 
-int image_pull([[maybe_unused]] const image_pull_args& args,
-               [[maybe_unused]] const global_settings& settings) {
+int image_pull(const image_pull_args& args, const global_settings& settings) {
     namespace fs = std::filesystem;
 
     if (args.build) {
@@ -86,8 +85,7 @@ int image_pull([[maybe_unused]] const image_pull_args& args,
         return 1;
     }
 
-    label.system =
-        site::get_system_name(label.system, settings.calling_environment);
+    label = apply_system(label, settings.config.system_name);
     if (!label.name) {
         term::error(
             "the uenv description '{}' must specify the name of the uenv",
@@ -131,14 +129,14 @@ int image_pull([[maybe_unused]] const image_pull_args& args,
     spdlog::info("pulling {} {}", record.sha, record);
 
     // require that a valid repo has been provided
-    if (!settings.config.repo) {
+    const auto repo = settings.config.repo();
+    if (!repo) {
         term::error("a repo needs to be provided either using the --repo "
                     "option, or in the config file");
         return 1;
     }
     // open the repo
-    auto store = uenv::open_repository(settings.config.repo.value(),
-                                       repo_mode::readwrite);
+    auto store = uenv::open_repository(repo->path, repo_mode::readwrite);
     if (!store) {
         term::error("unable to open repo: {}", store.error());
         return 1;