+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- #135 [fix] mount squasfs files on root squashed NFS mounts 
+- #133 [feature] improved error message for revoked registry tokens 
+- #131 [feature] remove --list flag in favor of a --format flag 
 - #129 [feature] `uenv image find` and `uenv image ls` do partial match on uenv names
 - #127 [feature] find views, mount point, etc. information using `uenv image inspect` without opening a uenv.